requirements.txt
Django==3.2.14
django-extensions==3.2.0
django-filter==22.1
djangorestframework==3.13.1
drf-yasg==1.21.3
psycopg2-binary==2.9.3
gunicorn==20.1.0
ipython==7.34.0
requests==2.28.1
